updating the code to only include paragraph tags
</p> tags seem to be where all the actual content is in the articles, though this excludes more sophisticated website designs, but whatever. At least this works well for article website formats.

It is still getting a bunch of text that is unnecessary for what we need but I can't think of a way to exclude this excess text without tailoring it for every specific website due to the div tags. If only we could just access an API for a json reponse of the text body. Oh well. Anyway this will work for now.