OOP refactor and implementing ChatGPT summaries
Writing out the code so that it runs the summaries from ChatGPT as an optional flag to summarize websites quicker for those who want to know the quick TLDR of the website.

I will improve upon this later, it is currently broken but I will fix it soon.

Otherwise, I redesigned the whole code from the ground up to include more OOP principles to make it a bit more standardized. and easily maintainable. :)