finished the summarizer with tgpt2
Unfortunately, the summaries are way off and don't make any sense or way over simplified to the point of not saying anything. So this would be perfect for generating job descriptions or corporate PR!!! :D

But sucks at summarizing websites.

Luckily, most of the code framework is modular and supports other AI models if we can get them to work. If we could get G4F to work with other models maybe it would be better. I'll try to inspect the code for tgpt2 because it looks like it's using g4f under the hood. Maybe I can see how it's calling it and reverse engineer that for use with gpt4 or gpt 3.5 or something to help out.

But for now, the proof concept is there.